захуярил отображение метрической высоты для пуссибоев, почему-то не любящих ножки

+++ main/UI/navigation/settings/spectrumScan/spectrumScanPage.cpp
@@ -283,8 +283,8 @@ namespace pizda {
 		const auto& bounds = getBounds();
 		const auto pointerPixelPos = pointerEventPos - bounds.getTopLeft();
 
-		_pointerPos.setX(std::clamp<float>(pointerPixelPos.getX() / bounds.getWidth(), 0, 1));
-		_pointerPos.setY(std::clamp<float>(1.f - pointerPixelPos.getY() / bounds.getHeight(), 0, 1));
+		_pointerPos.setX(std::clamp<float>(static_cast<float>(pointerPixelPos.getX()) / bounds.getWidth(), 0, 1));
+		_pointerPos.setY(std::clamp<float>(1.f - static_cast<float>(pointerPixelPos.getY()) / bounds.getHeight(), 0, 1));
 	}
 
 	SpectrumScanPage::SpectrumScanPage() {
